// src/app/model/user.ts
export class Users {
    id: number;
    first_name: string;
    last_name: string;
    date_on_born:any;
    id_address: number;
    telephone:number;
    username:string;
    roles:any;
    email:string;
    password:any;
    type_patient: number;
    sexe:string;
    token?: string;
    exp = 0;
}